requirements-lock.txt
alembic==1.6.4
amqp==5.0.6
appdirs==1.4.4
attrs==21.2.0
billiard==3.6.4.0
black==20.8b1
blinker==1.4
celery==5.1.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coverage==5.5
factory-boy==3.2.0
Faker==8.6.0
flake8==3.9.2
Flask==2.0.1
Flask-DB==0.3.1
Flask-DebugToolbar==0.11.0
Flask-Secrets==0.1.0
Flask-SQLAlchemy==2.5.1
Flask-Static-Digest==0.2.1
greenlet==1.1.0
gunicorn==20.1.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
kombu==5.1.0
Mako==1.1.4
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
prompt-toolkit==3.0.18
psycopg2==2.8.6
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
redis==3.5.3
regex==2021.4.4
six==1.16.0
SQLAlchemy==1.4.15
SQLAlchemy-Utils==0.37.4
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
